[fud2] Basic testing setup (#1954)
* Skeleton for the simplest possible test

* Reusable test scaffolding

* Test config setup

* Control executable name for tests

* Split main/lib/test

Now we have separate files for the main driver build, the actual `main`
function which is very short, and the tests.

* Add missing tests file

* Include through in test description

* A second test!

* Change config option name

We use `exe` everywhere, not `exec`, and I keep forgetting this fact.

* Be friendlier to Insta

By default, snapshot names are based on the call site... so let's try
separating the assertion lines?

* Factor out test function

* Accept snapshots!

* Add CI action to test fud2

* Appease clippy
