+++ fud2/Cargo.toml
@@ -18,6 +18,13 @@ anyhow.workspace = true
 manifest-dir-macros = "0.1"
 include_dir = "0.7"
 
+[lib]
+name = "fud2"
+path = "src/lib.rs"
+
 [[bin]]
 name = "fud2"
 path = "src/main.rs"
+
+[dev-dependencies]
+insta = "1.36.0"
