+++ fud2/fud-core/src/config.rs
@@ -15,6 +15,9 @@ pub struct GlobalConfig {
 
     /// Enable verbose output.
     pub verbose: bool,
+
+    /// The path to the build tool executable.
+    pub exe: String,
 }
 
 impl Default for GlobalConfig {
@@ -23,12 +26,17 @@ impl Default for GlobalConfig {
             ninja: "ninja".to_string(),
             keep_build_dir: false,
             verbose: false,
+            exe: std::env::current_exe()
+                .expect("executable path unknown")
+                .to_str()
+                .expect("invalid executable name")
+                .into(),
         }
     }
 }
 
 /// Location of the configuration file
-pub(crate) fn config_path(name: &str) -> std::path::PathBuf {
+pub fn config_path(name: &str) -> std::path::PathBuf {
     // The configuration is usually at `~/.config/driver_name.toml`.
     let config_base = env::var("XDG_CONFIG_HOME").unwrap_or_else(|_| {
         let home = env::var("HOME").expect("$HOME not set");
@@ -39,11 +47,15 @@ pub(crate) fn config_path(name: &str) -> std::path::PathBuf {
     config_path
 }
 
+/// Get raw configuration data with some default options.
+pub fn default_config() -> Figment {
+    Figment::from(Serialized::defaults(GlobalConfig::default()))
+}
+
 /// Load configuration data from the standard config file location.
-pub(crate) fn load_config(name: &str) -> Figment {
+pub fn load_config(name: &str) -> Figment {
     let config_path = config_path(name);
 
     // Use our defaults, overridden by the TOML config file.
-    Figment::from(Serialized::defaults(GlobalConfig::default()))
-        .merge(Toml::file(config_path))
+    default_config().merge(Toml::file(config_path))
 }
